\name{dataL}
\alias{dataL}
\docType{data}
\title{dataL
}
\description{
Dataset L used for examples.
}
\usage{data("dataL")}
\format{
  A data frame with 1000 observations on the following 56 variables.
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{x4}}{a numeric vector}
    \item{\code{x5}}{a numeric vector}
    \item{\code{x6}}{a numeric vector}
    \item{\code{x7}}{a numeric vector}
    \item{\code{x8}}{a numeric vector}
    \item{\code{x9}}{a numeric vector}
    \item{\code{x10}}{a numeric vector}
    \item{\code{x11}}{a numeric vector}
    \item{\code{x12}}{a numeric vector}
    \item{\code{x13}}{a numeric vector}
    \item{\code{x14}}{a numeric vector}
    \item{\code{x15}}{a numeric vector}
    \item{\code{x16}}{a numeric vector}
    \item{\code{x17}}{a numeric vector}
    \item{\code{x18}}{a numeric vector}
    \item{\code{x19}}{a numeric vector}
    \item{\code{x20}}{a numeric vector}
    \item{\code{x21}}{a numeric vector}
    \item{\code{x22}}{a numeric vector}
    \item{\code{x23}}{a numeric vector}
    \item{\code{x24}}{a numeric vector}
    \item{\code{x25}}{a numeric vector}
    \item{\code{x26}}{a numeric vector}
    \item{\code{x27}}{a numeric vector}
    \item{\code{x28}}{a numeric vector}
    \item{\code{x29}}{a numeric vector}
    \item{\code{x30}}{a numeric vector}
    \item{\code{x31}}{a numeric vector}
    \item{\code{x32}}{a numeric vector}
    \item{\code{x33}}{a numeric vector}
    \item{\code{x34}}{a numeric vector}
    \item{\code{x35}}{a numeric vector}
    \item{\code{x36}}{a numeric vector}
    \item{\code{x37}}{a numeric vector}
    \item{\code{x38}}{a numeric vector}
    \item{\code{x39}}{a numeric vector}
    \item{\code{x40}}{a numeric vector}
    \item{\code{x41}}{a numeric vector}
    \item{\code{x42}}{a numeric vector}
    \item{\code{x43}}{a numeric vector}
    \item{\code{x44}}{a numeric vector}
    \item{\code{x45}}{a numeric vector}
    \item{\code{x46}}{a numeric vector}
    \item{\code{x47}}{a numeric vector}
    \item{\code{x48}}{a numeric vector}
    \item{\code{x49}}{a numeric vector}
    \item{\code{x50}}{a numeric vector}
    \item{\code{x51}}{a numeric vector}
    \item{\code{x52}}{a numeric vector}
    \item{\code{x53}}{a numeric vector}
    \item{\code{x54}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
    \item{\code{npl}}{a numeric vector}
  }
}

\examples{
data(dataL)
## maybe str(dataL) ; plot(dataL) ...
}
\keyword{datasets}
